# backend/src/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "version": "1.0.3",
        "title": "Abacws Data Visualiser",
        "description": "Data visualing tool for the Cardiff University Comptuer Science Building (Abacws)",
        "license": {
            "name": "MIT"
        }
    },
    "servers": [
        {
            "url": "localhost:5000/api"
        }
    ],
    "paths": {
        "/devices": {
            "summary": "Get a list of all devices",
            "get": {
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/devices"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Unable to read/write from database",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/devices"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Unable to read/write from database",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/devices/{deviceName}": {
            "summary": "Get data for a device",
            "parameters": [
                { 
                    "$ref": "#/components/parameters/deviceName"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/deviceWithData"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Device doesn't exist",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Unable to read/write from database",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/deviceWithData"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Device doesn't exist",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Unable to read/write from database",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Add data to device",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Data accepted"
                    },
                    "503": {
                        "description": "Unable to read/write from database",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/devices/{deviceName}/history": {
            "summary": "Get historical data for a device",
            "parameters": [
                { 
                    "$ref": "#/components/parameters/deviceName"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/deviceWithHistory"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Device doesn't exist",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Unable to read/write from database",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/deviceWithHistory"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Device doesn't exist",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Unable to read/write from database",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Device doesn't exist",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Unable to read/write from database",
                        "content": {
                            "application/json" : {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "parameters": {
            "deviceName": { 
                "in": "path",
                "name": "deviceName",
                "required": true,
                "description": "Name of the the device to query",
                "schema": {
                    "type": "string",
                    "example": "Room-0.01"
                }
            }
        },
        "schemas": {
            "devices": {
                "type": "object",
                "properties": {
                    "offset": {
                        "type": "object",
                        "properties": {
                            "x": { "type": "number" },
                            "y": { "type": "number" },
                            "z": { "type": "number" }
                        }
                    },
                    "devices": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                }
            },
            "deviceWithData": {
                "type": "object",
                "properties": {
                    "name": { "type": "string" },
                    "data": {
                        "type": "object"
                    }
                }
            },
            "deviceWithHistory": {
                "type": "object",
                "properties": {
                    "name": { "type": "string" },
                    "history": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                }
            },
            "error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "examples": [
                            "Not Found",
                            "Could not read/write from database"
                        ]
                    }
                }
            }
        }
    }
}